backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-jwt-extended
python-dotenv
